Move card utils to `metabase/dashboard/utils`
